refactor(agent): extract linux/stub mods into separate files
agent/mod.rs reached 2420 lines with the entire Linux implementation
inline as `mod linux { ... }`. Move that body to `linux/mod.rs` and
`mod stub` to `stub.rs`, leaving mod.rs as a ~140-line shell of mod
declarations, the public entry point, and the tests module.

Pure mechanical move (dedent only). One incidental fix: making
`shared_containerd_config` `pub(super)` so the existing test access
through `super::linux::shared_containerd_config()` keeps compiling
now that the function lives in a child file. The same access on
master fails `cargo check --tests --target aarch64-unknown-linux-musl`
already; the move just surfaces the latent visibility bug.
